Adding a power meter with dual channel like this.
https://fr.aliexpress.com/item/1005007269856571.html

It's recognized as cz in Tuya , not powermeter (cz= switch or sensor).
I add the device creation from 19 to 22 (channel A & B, forward and reverse energy) and power storage. I use direction in order to set to 0 the other direction and avoid measurement error.
For example, if channel A is forward, it's reverse counter will be at 0.

For the moment this code is not running on my real server but in a sandbox server and it seems to work (but it's difficult to say if it's 100% good due to cloudy weather (few electricity produced by solar panes)).